sort_game/requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1
flasgger==0.9.7.1
gunicorn==21.2.0
PyJWT
bcrypt